Correcto arreglo del commit anterior

+++ scoring.py
@@ -128,7 +128,7 @@ class Scoring(object):
         # Calculo del idf para uso
         for dFreq in df:
             if dFreq is not 0:
-                idf.append(log10(long(N) / long(dFreq)))
+                idf.append(log10(float(N) / float(dFreq)))
             else:
                 idf.append(0)
         if mode: print "[debug]: IDF: %s" % idf
